requirements.txt
# LangChain and LangGraph dependencies
langchain>=0.1.0
langgraph>=0.0.1

# Environment variable management
python-dotenv>=1.0.0

# Optional - for development
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
